requirements.txt
scikit-learn
keras
tensorflow
pillow
opencv-python
djitellopy
numpy